feat: Move type, storage mode and version to modules
