+++ objects/src/accounts/account_id/mod.rs
@@ -2,7 +2,7 @@ mod id_anchor;
 pub use id_anchor::AccountIdAnchor;
 
 pub(crate) mod id_v0;
-pub use id_v0::{AccountIdV0, AccountIdVersion, AccountStorageMode, AccountType};
+pub use id_v0::AccountIdV0;
 
 mod id_prefix;
 pub use id_prefix::AccountIdPrefix;
@@ -14,3 +14,10 @@ mod id;
 pub use id::AccountId;
 
 mod seed;
+
+mod account_type;
+pub use account_type::AccountType;
+mod storage_mode;
+pub use storage_mode::AccountStorageMode;
+mod id_version;
+pub use id_version::AccountIdVersion;
